+++ unpythonic/ec.py
@@ -1,44 +1,85 @@
 # -*- coding: utf-8 -*-
-"""Escape continuations."""
+"""Escape continuations, i.e. jumping outward on the call stack.
 
-__all__ = ["escape", "setescape", "call_ec"]
+First see `call_ec`, which works like `call/ec` in Scheme and Racket. If you
+want more detailed control, see `catch` and `throw`.
 
+**Etymology**:
+
+The `catch`/`throw` functions are named after the `catch`/`throw` construct in
+Emacs Lisp and several precursors of Common Lisp, so this terminology dates
+from the mid-1980s, at the latest.
+
+Semantically this has nothing to do with exception handling, except that both
+are essentially a goto outward on the call stack. These names that are standard
+for this functionality in the Lisp family just happened to be available in
+Python, since Python's exception mechanism has chosen to use `except`/`raise`
+as its keywords.
+
+Common Lisp retains the `CATCH`/`THROW` construct, but there it is nowadays
+more idiomatic to use the lexically scoped variant `BLOCK`/`RETURN-FROM`
+(which we do not provide here).
+
+See Peter Seibel: Practical Common Lisp, chapter 20:
+    http://www.gigamonkeys.com/book/the-special-operators.html
+"""
+
+__all__ = ["throw", "catch", "call_ec",
+           "setescape", "escape"]  # old names, pre-0.14.2, will go away in 0.15.0
+
+from warnings import warn
 from functools import wraps
 
 from .regutil import register_decorator
 
 def escape(value, tag=None, allow_catchall=True):
-    """Escape to a ``@setescape`` point.
+    """Alias for `throw`, for backward compatibility.
+
+    Will be removed in 0.15.0.
+    """
+    warn("`escape` has been renamed `throw` as in Common Lisp; this alias will be removed in 0.15.0.", FutureWarning)
+    return throw(value, tag, allow_catchall)
+
+def setescape(tags=None, catch_untagged=True):
+    """Alias for `catch`, for backward compatibility.
+
+    Will be removed in 0.15.0.
+    """
+    warn("`setescape` has been renamed `catch` as in Common Lisp; this alias will be removed in 0.15.0.", FutureWarning)
+    return catch(tags, catch_untagged)
+
+def throw(value, tag=None, allow_catchall=True):
+    """Escape to a dynamically surrounding ``@catch``.
 
     Essentially this just raises an ``Escape`` instance with the given arguments.
     Wrapping the raise in a function call allows also lambdas to use escapes.
 
-    For the escape to work, we must be inside the dynamic extent
-    of the intended ``@setescape`` point.
+    For the ``throw`` to work, we must be inside the dynamic extent
+    of the intended ``@catch``.
 
     Parameters:
 
         value: anything
-            The value to send to the escape point.
+            The value to send to the ``@catch``.
 
         tag: anything comparable with ``==``
-            Can be used to restrict which ``@setescape`` points may catch
-            this escape instance. Default ``None``, meaning "no tag".
+            Can be used to restrict which ``@catch`` may catch
+            this `throw`. Default ``None``, meaning "no tag".
 
         allow_catchall: bool
-            Whether untagged "catch-all" ``@setescape`` points may catch
-            this escape instance (regardless of tag!).
+            Whether untagged "catch-all" ``@catch`` may catch
+            this `throw` (regardless of tag!).
 
-            Even if ``False``, if ``tag`` is ``None``, a ``@setescape``
+            Even if ``False``, if ``tag`` is ``None``, a ``@catch``
             point may still override this with ``catch_untagged=True``.
-            (See the table in ``help(setescape)``, case ``b0``.)
+            (See the table in ``help(catch)``, case ``b0``.)
     """
     raise Escape(value, tag, allow_catchall)
 
 class Escape(Exception):
-    """Exception that essentially represents an escape continuation.
+    """Exception that essentially represents the invocation of an escape continuation.
 
-    Constructor parameters: see ``escape()``.
+    Constructor parameters: see ``throw()``.
     """
     def __init__(self, value, tag=None, allow_catchall=True):
         self.value = value
@@ -46,13 +87,13 @@ class Escape(Exception):
         self.allow_catchall = allow_catchall
 
         # Error message when uncaught
-        self.args = ("Not within the dynamic extent of a @setescape",)
+        self.args = ("Not within the dynamic extent of a @catch",)
 
-def setescape(tags=None, catch_untagged=True):
-    """Decorator. Mark function as exitable by ``escape(value)``.
+def catch(tags=None, catch_untagged=True):
+    """Decorator. Mark function as exitable by ``throw(value)``.
 
     In Lisp terms, this essentially captures the escape continuation (ec)
-    of the decorated function. The ec can then be invoked by ``escape(value)``.
+    of the decorated function. The ec can then be invoked by ``throw(value)``.
 
     Technically, this is a decorator factory, since we take parameters.
 
@@ -61,14 +102,14 @@ def setescape(tags=None, catch_untagged=True):
           where ``t`` is anything comparable with ``==``
             A tuple is OR'd, like in ``isinstance()``.
 
-            Restrict which escapes will be caught by this ``@setescape`` point.
+            Restrict which ``throw`` will be caught by this ``@catch``.
 
         catch_untagged: bool
-            Whether this ``@setescape`` point catches untagged escapes.
+            Whether this ``@catch`` catches untagged throws.
 
-            Even if ``False``, if ``tags`` is ``None``, an escape instance
-            may still override this with ``allow_catchall=True``. (See the
-            table below, case ``a1``.)
+            Even if ``False``, if ``tags`` is ``None``, a ``throw``
+            may still override this with ``allow_catchall=True``.
+            (See the table below, case ``a1``.)
 
     The exact catch condition is::
 
@@ -78,10 +119,10 @@ def setescape(tags=None, catch_untagged=True):
            (tags is not None and e.tag is not None and e.tag in tags):
             # caught
 
-    resulting in the following table. Here a ``@setescape`` point with no tags
+    resulting in the following table. Here a ``@catch`` point with no tags
     is a "catch-all"::
 
-        escape instance                     @setescape point
+        throw instance                      @catch point
         0: no tag, ignore catch-alls        a: no tags
         1: no tag, allow catch-alls         b: no tags, catch untagged
         2: w/ tag, ignore catch-alls        c: w/ tags
@@ -99,9 +140,9 @@ def setescape(tags=None, catch_untagged=True):
 
     **How to use the table**:
 
-      - If setting a ``@setescape`` point, pick a row. Read off what you'll catch.
+      - If setting a ``@catch``, pick a row. Read off what you'll catch.
 
-      - If raising an ``escape`` instance, pick a column. Read off what will catch it.
+      - If raising a ``throw`` instance, pick a column. Read off what will catch it.
 
     Default settings for both ends give case ``b1``. For a guaranteed one-to-one
     relationship, pick a unique tag, and use settings for case ``c2``.
@@ -110,10 +151,10 @@ def setescape(tags=None, catch_untagged=True):
 
     Multi-return using escape continuation::
 
-        @setescape()
+        @catch()
         def f():
             def g():
-                escape("hello from g")  # the arg becomes the return value of f()
+                throw("hello from g")  # the arg becomes the return value of f()
                 print("not reached")
                 return False
             g()
@@ -123,28 +164,28 @@ def setescape(tags=None, catch_untagged=True):
 
     Escape from FP loop::
 
-        @setescape()  # no tag, catch any escape instance
+        @catch()  # no tag, catch any throw
         def f():
             @looped
             def s(loop, acc=0, i=0):
                 if i > 5:
-                    escape(acc)  # the argument becomes the return value of f()
+                    throw(acc)  # the argument becomes the return value of f()
                 return loop(acc + i, i + 1)
             print("never reached")
             return False
         assert f() == 15
 
-    For more control, tagged escape::
+    For more control, tagged throws::
 
-        @setescape("foo")
+        @catch("foo")
         def foo():
             @call
-            @setescape("bar")
+            @catch("bar")
             def bar():
                 @looped
                 def s(loop, acc=0, i=0):
                     if i > 5:
-                        escape(acc, tag="foo")
+                        throw(acc, tag="foo")
                     return loop(acc + i, i + 1)
                 print("never reached")
                 return False
@@ -155,12 +196,12 @@ def setescape(tags=None, catch_untagged=True):
     if tags is not None:
         if isinstance(tags, tuple):  # multiple tags
             tags = set(tags)
-        else: # single tag
+        else:  # single tag
             tags = set((tags,))
 
     def decorator(f):
         @wraps(f)
-        def escapepoint(*args, **kwargs):
+        def catchpoint(*args, **kwargs):
             try:
                 return f(*args, **kwargs)
             except Escape as e:
@@ -170,7 +211,7 @@ def setescape(tags=None, catch_untagged=True):
                     return e.value
                 else:  # meant for someone else, pass it on
                     raise
-        return escapepoint
+        return catchpoint
     return decorator
 
 @register_decorator(priority=80)
@@ -183,14 +224,14 @@ def call_ec(f):
             the first-class escape continuation (ec).
 
             The ec, in turn, takes one positional argument; the value
-            to send to the escape point. It becomes the return value
+            to send to the catch point. It becomes the return value
             of the ``call_ec``.
 
-            The ec instance and the escape point are connected one-to-one.
+            The ec instance and the catch point are connected one-to-one.
             Both are tagged with a process-wide unique id. The ec is set to
             disallow catch-all (so that it will always reach its intended
-            destination), and the point is set to ignore any untagged escapes
-            (so that it catches only this particular ec).
+            destination), and the catch point is set to ignore any untagged
+            throws (so that it catches only this particular ec).
 
     Like in ``@call``, the function ``f`` is called immediately,
     and the def'd name is replaced by the return value of ``f(ec).``
@@ -201,7 +242,7 @@ def call_ec(f):
         def result(ec):
             answer = 42
             def inner():
-                ec(answer)  # directly escapes from the outer def
+                ec(answer)  # directly escape from the outer def
                 print("never reached")
                 return 23
             answer = inner()
@@ -236,11 +277,11 @@ def call_ec(f):
     def ec(value):
         if not ec_valid:
             raise RuntimeError("Cannot escape after the dynamic extent of the call_ec invocation.")
-        # Be catchable only by our own escape point.
-        escape(value, uid, allow_catchall=False)
+        # Be catchable only by our own catch point.
+        throw(value, uid, allow_catchall=False)
     try:
-        # Set up a tagged escape point that catches only the ec we just set up.
-        @setescape(uid, catch_untagged=False)
+        # Set up a tagged catch point that catches only the ec we just set up.
+        @catch(uid, catch_untagged=False)
         def wrapper():
             return f(ec)
         return wrapper()